requirements.txt
streamlit
qdrant-client
openai
cohere
tenacity
python-dotenv
pandas
